+++ app/components/MarketFilter/marketFilter.module.scss
@@ -1,0 +1,293 @@
+
+@keyframes slideDown {
+  from {
+    max-height: 0;
+  }
+
+  to {
+    max-height: 100%;
+  }
+}
+
+@keyframes slideUp {
+  from {
+    max-height: 100%;
+  }
+
+  to {
+    max-height: 0;
+  }
+}
+
+.filterContainer {
+  max-height: 58px;
+  transition: max-height 0.3s ease-in;
+}
+
+.filterContainer_active {
+  max-height: calc(100% - 270px);;
+}
+
+.filters {
+  max-height: 0;
+  margin: 10px 10px 0 10px;
+  overflow: hidden;
+  transition: max-height 0.3s ease-in;
+}
+
+.filters__active {
+  animation: slideDown 0.3s forwards
+}
+
+.filters:not(.filters__active) {
+  animation: slideUp 0.3s forwards;
+}
+
+.filterContainer {
+  display: flex;
+  flex-direction: column;
+  background: #3a4340;
+  border-radius: 10px 10px 0 0;
+
+  &__active {
+    height: calc(100% - 270px);
+  }
+}
+
+.titleBorder {
+  width: 100%;
+  min-height: 58px;
+  padding: 1px;
+  background: #474c52;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  .filterTitle {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    height: 100%;
+    background: #1c1c1c;
+    color: #d9d9d9;
+
+    &:active {
+      background: #0f629a;
+    }
+
+    h3 {
+      font-size: 16px;
+      font-weight: 700;
+    }
+
+    &__active {
+      border-radius: 0 10px;
+      color: #42b5ff;
+    }
+  }
+
+  &__active {
+    border-radius: 0 10px;
+    background: linear-gradient(90deg, rgba(24, 134, 208, 1) 0%, rgba(12, 68, 106, 1) 100%);
+  }
+
+  &:hover {
+    background: linear-gradient(90deg, rgba(24, 134, 208, 1) 0%, rgba(12, 68, 106, 1) 100%);
+  }
+}
+
+.filterBox {
+  box-shadow: inset 4px 4px 20px 0 rgba(4, 4, 4, 0.25);
+  background: #212121;
+  border-radius: 10px 10px 0 0;
+  padding: 10px;
+  color: #D9D9D9;
+  height: calc(100% - 10px);
+  overflow-y: scroll;
+
+  &::-webkit-scrollbar {
+    width: 8px;
+    background: none;
+  }
+
+  &::-webkit-scrollbar-track {
+    background: none;
+  }
+
+  &::-webkit-scrollbar-thumb {
+    background: #d9d9d9;
+    border-radius: 10px;
+    cursor: pointer;
+  }
+
+  &::-webkit-scrollbar-thumb:hover {
+    background: #555;
+  }
+}
+
+.resetBtn {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 10px;
+
+  button {
+    border-radius: 10px;
+    width: 215px;
+    height: 39px;
+    padding: 10;
+    color: #d9d9d9;
+    background: #1c1c1c;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 0 20px;
+
+    &:hover {
+      background: #3a4340;
+    }
+
+    &:active {
+      background: #1c1c1c;
+      border: 1px solid #0f629a;
+    }
+
+    span {
+      font-weight: 700;
+      font-size: 18px;
+    }
+
+    .minusSymbol {
+      color: #1886D0;
+      font-weight: 700;
+      font-size: 24px;
+    }
+  }
+}
+
+.dropdownBox {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.antInputNumber {
+  border: 1px solid #3a4340;
+  border-radius: 5px;
+  width: 79px;
+  height: 27px;
+}
+
+.sliderInputsBox {
+  display: flex;
+  margin-bottom: 12px;
+
+  b {
+    margin: 0 10px;
+  }
+}
+
+.sliderBox {
+  margin-bottom: 20px;
+  padding: 0 5px;
+}
+
+.filterInput {
+  color: #d9d9d9;
+  background: #1c1c1c;
+
+  &:hover {
+    background: #3a4340;
+  }
+
+  &:focus {
+    background: #1c1c1c;
+    border: 1px solid #0f629a;
+  }
+}
+
+.deliveryInputBox {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 10px;
+}
+
+.checkboxContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: start;
+  margin-bottom: 10px;
+}
+
+.customCheckbox {
+  display: flex;
+  align-items: center;
+  margin-bottom: 10px;
+
+  input:checked+.checkbox {
+    border: 1px solid #0f629a;
+
+    .checkedIcon {
+      display: block;
+    }
+  }
+}
+
+.checkbox {
+  font-weight: 400;
+  font-size: 14px;
+  border-radius: 5px;
+  width: 30px;
+  height: 30px;
+  border: 1px solid #3a4340;
+  background: #181819;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  margin-right: 10px;
+}
+
+.hiddenCheckbox {
+  appearance: none;
+  -webkit-appearance: none;
+  position: absolute;
+}
+
+.checkbox .checkedIcon {
+  max-width: 67%;
+  height: auto;
+  display: none;
+}
+
+.colorList {
+  border-radius: 10px;
+  width: 215px;
+  height: 39px;
+  background: transparent;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 0 20px;
+  font-weight: 600;
+  border: 0;
+  color: #d9d9d9;
+  
+
+  .btnName {
+    font-size: 18px;
+  }
+
+  .plusSymbol {
+    font-size: 24px;
+  }
+}
+
+.ant-btn {
+  &:hover {
+    background: red;
+  }
+}
